separate read node spawn
