+++ src/consensus/malachite/mod.rs
@@ -4,6 +4,7 @@ pub mod read_host;
 pub mod read_sync;
 pub mod snapchain_codec;
 pub mod spawn;
+pub mod spawn_read_node;
 
 #[cfg(test)]
 mod read_node_actors_test;
