+++ src/consensus/malachite/spawn.rs
@@ -5,7 +5,6 @@ use informalsystems_malachitebft_engine::host::HostRef;
 use informalsystems_malachitebft_engine::network::NetworkRef;
 use informalsystems_malachitebft_network::{PeerId as MalachitePeerId, PeerIdExt};
 use informalsystems_malachitebft_sync::Metrics as SyncMetrics;
-use std::collections::BTreeMap;
 use std::path::Path;
 use tracing::Span;
 
@@ -16,20 +15,15 @@ use crate::consensus::malachite::network_connector::{
     NetworkConnectorArgs,
 };
 use crate::consensus::malachite::snapchain_codec::SnapchainCodec;
-use crate::consensus::read_validator::{self, Engine};
 use crate::consensus::validator::ShardValidator;
 use crate::core::types::{ShardId, SnapchainValidatorContext};
 use crate::network::gossip::GossipEvent;
-use crate::proto::{self, Height};
 use informalsystems_malachitebft_engine::sync::{Params as SyncParams, Sync, SyncRef};
 use informalsystems_malachitebft_engine::util::events::TxEvent;
 use informalsystems_malachitebft_engine::wal::{Wal, WalRef};
 use informalsystems_malachitebft_metrics::{Metrics, SharedRegistry};
 use libp2p::PeerId;
 use tokio::sync::mpsc;
-
-use super::read_host::{ReadHost, ReadHostMsg, ReadHostRef, ReadHostState};
-use super::read_sync::{ReadParams, ReadSync, ReadSyncRef};
 
 pub async fn spawn_network_actor(
     gossip_tx: mpsc::Sender<GossipEvent<SnapchainValidatorContext>>,
@@ -75,26 +69,6 @@ pub async fn spawn_host(
         gossip_tx,
     };
     let actor_ref = Host::spawn(state).await?;
-    Ok(actor_ref)
-}
-
-pub async fn spawn_read_host(
-    shard_id: u32,
-    engine: Engine,
-) -> Result<ReadHostRef, ractor::SpawnErr> {
-    let state = ReadHostState {
-        validator: read_validator::ReadValidator {
-            shard_id,
-            engine,
-            last_height: Height {
-                shard_index: shard_id,
-                block_number: 0,
-            },
-            max_num_buffered_blocks: 100,
-            buffered_blocks: BTreeMap::new(),
-        },
-    };
-    let actor_ref = ReadHost::spawn(state).await?;
     Ok(actor_ref)
 }
 
@@ -156,99 +130,6 @@ pub async fn spawn_sync_actor(
     let actor_ref = Sync::spawn(ctx, network, host, params, metrics, span).await?;
 
     Ok(actor_ref)
-}
-
-pub async fn spawn_read_sync_actor(
-    ctx: SnapchainValidatorContext,
-    network: NetworkRef<SnapchainValidatorContext>,
-    host: ReadHostRef,
-    config: SyncConfig,
-    registry: &SharedRegistry,
-    span: Span,
-) -> Result<ReadSyncRef, ractor::SpawnErr> {
-    let params = ReadParams {
-        status_update_interval: config.status_update_interval,
-        request_timeout: config.request_timeout,
-    };
-
-    let metrics = SyncMetrics::register(registry);
-
-    let actor_ref = ReadSync::spawn(ctx, network, host, params, metrics, span).await?;
-
-    Ok(actor_ref)
-}
-
-#[derive(Clone)]
-pub struct MalachiteReadNodeActors {
-    pub network_actor: NetworkRef<SnapchainValidatorContext>,
-    pub host_actor: ReadHostRef,
-    pub sync_actor: ReadSyncRef,
-}
-
-impl MalachiteReadNodeActors {
-    pub async fn create_and_start(
-        ctx: SnapchainValidatorContext,
-        engine: Engine,
-        local_peer_id: PeerId,
-        gossip_tx: mpsc::Sender<GossipEvent<SnapchainValidatorContext>>,
-        registry: &SharedRegistry,
-        shard_id: u32,
-    ) -> Result<Self, ractor::SpawnErr> {
-        let name = if shard_id == 0 {
-            format!("Block")
-        } else {
-            format!("Shard {}", shard_id)
-        };
-        let span = tracing::info_span!("node", name = %name);
-
-        let network_actor = spawn_network_actor(gossip_tx.clone(), local_peer_id).await?;
-        let host_actor = spawn_read_host(shard_id, engine).await?;
-        let sync_actor = spawn_read_sync_actor(
-            ctx.clone(),
-            network_actor.clone(),
-            host_actor.clone(),
-            SyncConfig::default(),
-            registry,
-            span.clone(),
-        )
-        .await?;
-
-        host_actor
-            .cast(ReadHostMsg::Started {
-                sync: sync_actor.clone(),
-            })
-            .unwrap();
-
-        Ok(Self {
-            network_actor,
-            host_actor,
-            sync_actor,
-        })
-    }
-
-    pub fn cast_decided_value(
-        &self,
-        value: proto::DecidedValue,
-    ) -> Result<(), ractor::MessagingErr<ReadHostMsg>> {
-        self.host_actor.cast(ReadHostMsg::ProcessDecidedValue {
-            value,
-            sync: self.sync_actor.clone(),
-        })
-    }
-
-    pub fn cast_network_event(
-        &self,
-        event: MalachiteNetworkEvent,
-    ) -> Result<(), ractor::MessagingErr<MalachiteNetworkActorMsg>> {
-        self.network_actor
-            .cast(MalachiteNetworkActorMsg::NewEvent(event))
-    }
-
-    pub fn stop(&self) {
-        self.host_actor.stop(None);
-        self.network_actor.stop(None);
-        self.sync_actor.stop(None);
-    }
 }
 
 #[derive(Clone)]
