+++ src/consensus/read_validator_test.rs
@@ -4,7 +4,7 @@ mod tests {
     use std::collections::BTreeMap;
 
     use crate::consensus::malachite::read_sync::ReadSyncRef;
-    use crate::consensus::malachite::spawn::MalachiteReadNodeActors;
+    use crate::consensus::malachite::spawn_read_node::MalachiteReadNodeActors;
     use crate::consensus::read_validator::{Engine, ReadValidator};
     use crate::core::types::SnapchainValidatorContext;
     use crate::network::gossip::GossipEvent;
