+++ src/node/snapchain_read_node.rs
@@ -1,6 +1,6 @@
 use crate::consensus::consensus::{Config, MalachiteEventShard};
 use crate::consensus::malachite::network_connector::MalachiteNetworkEvent;
-use crate::consensus::malachite::spawn::MalachiteReadNodeActors;
+use crate::consensus::malachite::spawn_read_node::MalachiteReadNodeActors;
 use crate::consensus::read_validator::Engine;
 use crate::core::types::{Address, ShardId, SnapchainShard, SnapchainValidatorContext};
 use crate::mempool::mempool::MempoolMessagesRequest;
